moved recorder and reader out
